requirements.txt
Django==3.2.8
gunicorn==20.1.0
psycopg2-binary>=2.8
whitenoise==5.3.0
Pillow==8.4.0
django-redis==5.2.0
django-cors-headers==3.11.0
django-ninja==0.17.0
pydantic==1.9.0
